refactor(customers): decompose CustomersService into Write/Query/Analytics services
Behavior-preserving. Facade keeps the 17-method surface + delegates. The 2
$transaction write paths (create, findOrCreatePrecheckCustomer) moved WHOLE into
CustomerWriteService; read+decrypt into CustomerQueryService; $queryRaw reports
into CustomerAnalyticsService. @Optional piiService + inline PII fallback kept
verbatim (not made required). tests green, tsc 0, lint 0.
